dm: don't send uevents while the device is suspended
Device mapper sends an uevent when the device is suspended, using the
function set_capacity_and_notify. However, this causes a race condition
with udev.

Udev skips scanning dm devices that are suspended. If we send an uevent
while we are suspended, udev will be racing with device mapper resume
code. If the device mapper resume code wins the race, udev will process
the uevent after the device is resumed and it will properly scan the
device.

However, if udev wins the race, it will receive the uevent, find out that
the dm device is suspended and skip scanning the device. This causes bugs
such as systemd unmounting the device - see
https://bugzilla.redhat.com/show_bug.cgi?id=2158628

This commit fixes this race.

We move the code that sends the uevents from __bind to dm_resume, so that
they are sent just before the dm-device is resumed. We hold
md->suspend_lock while sending the uevent. We change __dev_status so that
it grabs md->suspend_lock while querying the suspend state.

If udev responds too quickly after the uevent is generated, it will go to
__dev_status, call dm_suspended_md_locked and that function will block
until md->suspend_lock is released by the resume code - so, udev won't see
the device in the suspended state.

Signed-off-by: Mikulas Patocka <mpatocka@redhat.com>
Tested-by: Peter Rajnoha <prajnoha@redhat.com>
Cc: stable@vger.kernel.org
Signed-off-by: Mike Snitzer <snitzer@kernel.org>