arm64: dts: qcom: sdm845: add framebuffer reserved memory
The bootloader configures a framebuffer in memory to display splash
screens or other information. This configuration is overriden when the
display pipeline probes, but never unmapped from the SMMU. To prevent
issues from the kernel trying to allocate in this region and to allow
using the framebuffer for debugging, many devices already reserve this
region.

All devices follow MTP and use the same address for this region, Cheza
is a likely exception though I'm not able to validate that.

Some devices only reserve the size needed to store the actual
framebuffer, this is incorrect as on all devices I've checked the full
0x2400000 bytes are mapped.

This patch moves the framebuffer region to sdm845.dtsi and removes it
from each individual device. This ensures that the correct size is
always reserved and prevents having the add the region for each
individual device.

This patch specifically ensures that this region is now reserved on the
OnePlus 6, where it wasn't before.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>
Link: https://lore.kernel.org/r/20230315111947.3807083-1-caleb.connolly@linaro.org