arm64: dts: qcom: sdm850: commonise framebuffer memory
This region needs to be reserved as it's mapped by the bootloader and
the kernel doesn't currently have a way of unmapping it.

Both the Galaxy Book2 and C630 use the same address and size for the
framebuffer reserved memory region. Move the definition to sdm850.dtsi.

The c630 was missing this region, this patch ensures that it's reserved
there too.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>
Link: https://lore.kernel.org/r/20230315111947.3807083-2-caleb.connolly@linaro.org