media: venus: hfi_venus: Don't wait for PC_PREP_DONE on AR50L/IRIS2
AR50_LITE will never send a power collapse ready ack, so don't wait
for it to avoid timeouts in the middle of the hw suspend sequence.
We can rely on the following venus_cpu_idle_and_pc_ready check instead.

IRIS2 seems to not have that issue on SM8250, but it does manifest on
SM8350, so do the same for it.

Signed-off-by: Konrad Dybcio <konrad.dybcio@linaro.org>
Signed-off-by: Johan Hovold <johan+linaro@kernel.org>