i2c: amd-mp2: Remove #ifdef guards for PM related functions
Use the new PM macros for the suspend and resume functions to be
automatically dropped by the compiler when CONFIG_PM or
CONFIG_PM_SLEEP are disabled, without having to use #ifdef guards.

This has the advantage of always compiling these functions in,
independently of any Kconfig option. Thanks to that, bugs and other
regressions are subsequently easier to catch.

Note that the use of the UNIVERSAL_DEV_PM_OPS() macro was likely to be
wrong, as it sets the same callbacks for the runtime-PM and system
suspend/resume. This patch does not change this behaviour, but I suspect
that it should be changed to use DEFINE_RUNTIME_DEV_PM_OPS() instead, as
the current documentation for UNIVERSAL_DEV_PM_OPS() suggests.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>
Link: https://lore.kernel.org/r/20230705204314.89800-2-paul@crapouillou.net