soc: qcom: add pd-mapper implementation
The existing userspace protection domain mapper implementation has
several issue. It doesn't play well with CONFIG_EXTRA_FIRMWARE, it
doesn't reread the JSON files if the firmware location is changed (or if
the firmware was not available at the time pd-mapper was started but the
corresponding directory is mounted later), etc.

Provide in-kernel service implementing protection domain mapping
required to work with several services, which are provided by the DSP
firmware.

Signed-off-by: Dmitry Baryshkov <dmitry.baryshkov@linaro.org>
Link: https://lore.kernel.org/r/20231226003447.3044365-2-dmitry.baryshkov@linaro.org