media: venus: fix OOB read issue due to double read
During message queue read, the address is being read twice
from the shared memory. The first read is validated against
the size of the packet, however the second read is not
being validated. Therefore, it's possible for firmware to
modify the value to a bigger invalid value which can lead
to OOB read access issue while reading the packet.
Added fix to reupdate the size of the packet which was
read for the first time.

Signed-off-by: Vedang Nagar <quic_vnagar@quicinc.com>
Link: https://lore.kernel.org/r/20250104-venus-security-fixes-v1-1-9d0dd4594cb4@quicinc.com