bus: mhi: host: Allocate entire MHI control config once
MHI control configurations such as channel context, event context, command
context and rings, are currently allocated individually. During MHI
initialization MHI bus driver needs to configure the address space in
which this control configuration resides. Since different component of the
config is being allocated separately, only logical solution is to give the
entire RAM address space, as they could be anywhere.

According to the MHI specification v1.2 rev AB the MHI control
configuration address space should not be more them 4GB. This comes from
the description of MHICTRLLIMIT in section 6.2 which states: "The most
significant 32 bits of MHICTRLBASE and MHICTRLLIMIT registers must be
equal".

This means we have a 4GB range (32-bit) to play with in a 64-bit address
space.  If the upper 32-bits of the 64-bit address for both the base and
the limit must be the same, then the range of addresses from the base to
the limit can only vary the lower 32-bits.

Invalid:
BASE: 0x0
LIMIT: 0xffffffff_ffffffff

Valid:
BASE: 0x0f_00000000
LIMIT: 0x0f_ffffffff

The current implementation is in violation of MHI specification.

For example mhi_init_dev_ctxt() -

We allocate the chan_ctxt with dma_alloc_coherent() as an individual
allocation.  In the case of AIC100, the device can access the full
64-bit address space, but the DMA engine is limited to a 32-bit transfer
size.  The chan_ctxt probably won't be larger than 4GB, so the size is
rather irrelevant.  Can be allocated anywhere.  Lets say that it gets
put in the lower 32-bit address space - 0x0_XXXXXXXX

Then a little bit later we allocate er_ctxt with a different
dma_alloc_coherent() instance.  Being a unique allocation, it is not
tied to the chan_ctxt and can exist anywhere.  Lets assume that it gets
put somewhere in the non-lower 32-bits - 0x1000_XXXXXXXX

Now we have a problem because we cannot describe a single range covering
both of these allocations via MHICTRLBASE/MHICTRLLIMIT where the upper
32-bits of both registers is the same.

Change the implementation to comply with the MHI spec by allocating a
single giant DMA buffer for MHI control configurations and limit the
configuration address space to that buffer.

Signed-off-by: Pranjal Ramajor Asha Kanojiya <quic_pkanojiy@quicinc.com>
Signed-off-by: Jeff Hugo <jeff.hugo@oss.qualcomm.com>
Reviewed-by: Manivannan Sadhasivam <manivannan.sadhasivam@linaro.org>
Link: https://lore.kernel.org/r/20250425170431.921656-1-jeff.hugo@oss.qualcomm.com