drm/msm/a6xx: Poll AHB fence status in GPU IRQ handler
Even though GX power domain is kept ON when there is a pending GPU
interrupt, there is a small window of potential race with GMU where it
may move the AHB fence to 'Drop' mode. Close this race window by polling
for AHB fence to ensure that it is in 'Allow' mode.

Signed-off-by: Akhil P Oommen <akhilpo@oss.qualcomm.com>
Link: https://lore.kernel.org/r/20250720-ifpc-support-v1-10-9347aa5bcbd6@oss.qualcomm.com