remoteproc: qcom_q6v5: Add qcom_q6v5_attach()
Implement qcom_q6v5_attach(), which can be used by q6v5 remoteproc drivers
to attach to a remoteproc that is already running during boot. It is
essentially the same as qcom_q6v5_prepare(), except that we skip setting up
the proxy votes, since we expect that handover has already happened.

Signed-off-by: Stephan Gerhold <stephan.gerhold@linaro.org>