net: qrtr: smd: Fix race conditions while opening/closing channel
If rpmsg drivers specify a receive callback in the rpmsg_driver struct, the
rpmsg core creates the endpoint (opens the channel) before calling the
driver probe() function, and similarly it destroys the endpoint (closes the
channel) after calling the driver remove() function. The receive callback
can be invoked from interrupt context any time when the channel is open,
before (or in parallel) of the probe() function, and after (or in parallel)
of the remove() function. Only few rpmsg drivers handle that reliably.

The QRTR SMD driver partially attempts to handle one of these edge cases by
checking if dev_get_drvdata() is already set when the receive callback is
invoked. If not, it returns -EAGAIN in the hope that the receive callback
will be repeated at a later point in time. Unfortunately, this can still
lead to problems:

 - The qcom_smd rpmsg backend (somewhat) handles -EAGAIN because it
   re-checks the channel status later during announce_create(), but the
   glink smem backend simply ignores that. The rpmsg backend cannot really
   know when "again" is supposed to be - during the next incoming
   interrupt, in a few seconds, or in a few years?

 - There are no memory barriers, so just because the driver data (set at
   the end of the probe() function) already became visible does not mean
   the whole struct has already become visible to another CPU.

 - During remove(), the driver data is reset to NULL after the endpoint is
   already unregistered. Packets could still come inbetween that.

One rare issue that has been reported occasionally over the years is that
the ADSP on Qualcomm compute platforms (e.g. X13s or X1E laptops) does not
always come up reliably (depending on kernel configuration and timing). One
particular symptom is that the ADSP appears to function correctly (e.g.
charging LED changes state), but no QMI services are exposed via QRTR. This
can happen when the initial QRTR "hello" packet gets dropped, which is sent
by the remoteproc immediately upon opening the channel.

In order to fix this issue properly, it is easiest to inverse the startup
sequence and register the QRTR endpoint before opening the rpmsg channel.
This is because we can easily block on a mutex lock in the send path, but
could only use a short-running spin lock in the receive callback:

 - Register the QRTR endpoint before opening the channel (and unregister
   after closing the channel), so that we can always handle incoming
   packets.

 - Block the send path with a lock until we have fully initialized.

 - During remove(), start rejecting all further send operations with
   -ENODEV before we close the rpmsg channel.

 - The startup ordering and the lock implicitly provide the necessary
   memory barriers.

Use the new rpmsg_dev_open_ept() function to have more control over the
initialization order and add the necessary locking.

It is difficult to identify a proper Fixes tag. The qcom_smd rpmsg backend
that this driver was originally intended for (hence the name qrtr/smd.c)
handles -EAGAIN somewhat, but the newer glink smem backend does not, so
this is where things started (rarely) breaking.

Cc: stable@vger.kernel.org
Fixes: caf989c350e8 ("rpmsg: glink: Introduce glink smem based transport")
Signed-off-by: Stephan Gerhold <stephan.gerhold@linaro.org>